test: add SearchEntityConstructorArgsMustBeNamedTest, extract shared parsing support
Now that the read-side (io.camunda.search.entities) retrofit above closes
the same gap the write-side RdbmsDbModelConstructorArgsMustBeNamedTest
already enforces, add a sibling test for it instead of leaving the read
side unprotected against future regressions.

Extracted the near-identical parse-and-check logic (parse all 38 mapper
XMLs into a real MyBatis Configuration; assert every constructor arg has
name=; assert no statement relies on implicit resultType= automapping)
into a shared, package-private MapperConfigurationTestSupport, parameterized
by package prefix, rather than duplicating ~90 lines across two files.
RdbmsDbModelConstructorArgsMustBeNamedTest is refactored to call it --
behavior-preserving, same assertions and messages as before.

Both checks sabotage-tested directly for the new read-side prefix: a
name= typo (via the shared write-side test's typo, which the parser
catches first, then isolated to only the read-side occurrence) and a
reverted resultType= fix both correctly fail; restored afterward.
